fix(desktop): address storage growth review feedback
