+++ apps/desktop/src/renderer/lib/persisted-keys/index.ts
@@ -1,7 +1,1 @@
-export {
-	DEAD_KEYS,
-	LIBRARY_OWNED_KEY_PREFIXES,
-	PERSISTED_KEY_REGISTRY,
-	type PersistedKeyOwner,
-	sweepDeadPersistedKeys,
-} from "./persisted-keys";
+export { sweepDeadPersistedKeys } from "./persisted-keys";
